requirements.txt
accelerate==0.31.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
einops==0.8.0
exceptiongroup==1.2.2
filelock==3.14.0
flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
# google-ai-generativelanguage==0.4.0
# google-api-core==2.19.1
# google-auth==2.34.0
google-generativeai==0.4.1
# googleapis-common-protos==1.63.2
greenlet==3.0.3
# grpcio==1.65.5
# grpcio-tools==1.65.5
# grpcio-status==1.48.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
importlib-metadata==7.1.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-text-splitters==0.2.1
langsmith==0.1.75
llama-cloud==0.0.13
llama-index==0.10.67.post1
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.67
llama-index-embeddings-huggingface==0.2.1
llama-index-embeddings-langchain==0.1.2
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48.post3
llama-index-llms-huggingface==0.2.3
llama-index-llms-ollama==0.1.5
llama-index-llms-openai==0.1.29
llama-index-llms-replicate==0.1.3
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
llamaindex-py-client==0.1.19
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
minijinja==2.0.1
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.4.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
# numpy==1.17
openai==1.41.1
orjson==3.10.3
# packaging==24.1
pandas==2.2.2
pillow==10.4.0
pip==20.2.3
# proto-plus==1.24.0
# protobuf==3.19.5
psutil==5.9.8
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.8.2
pydantic-core==2.20.1
pydeck==0.9.1
pygments==2.18.0
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
setuptools==49.2.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
# streamlit==1.37.1
striprtf==0.0.26
sympy==1.12.1
tbb==2021.12.0
tenacity==8.5.0
text-generation==0.7.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
torch==2.3.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.41.2
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2
werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
